doctor: move code away from cmd/
