+++ Library/Homebrew/cmd/install.rb
@@ -1,5 +1,5 @@
 require "blacklist"
-require "cmd/doctor"
+require "diagnostic"
 require "cmd/search"
 require "formula_installer"
 require "tap"
@@ -159,7 +159,7 @@ module Homebrew
   end
 
   def check_xcode
-    checks = Checks.new
+    checks = Diagnostic::Checks.new
     %w[
       check_for_unsupported_osx
       check_for_bad_install_name_tool
